uart-mcp/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "uart-mcp"
version = "0.1.0"
description = "Bidirectional UART MCP server — session-based serial console interaction"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "pyserial>=3.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
uart-mcp = "uart_mcp.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
